+++ tests/testthat/test_regr_xgboost.R
@@ -228,18 +228,27 @@ test_that("mlr3measures are equal to internal measures", {
 })
 
 test_that("base_margin (offset)", {
-  task = tsk("mtcars")
+  data = with_seed(1, {
+    n = 200
+    x = rnorm(n)
+    eta = 0.3 * x
+    lambda = exp(eta)
+    y = rpois(n, lambda)
+    data.frame(x, y)
+  })
 
-  # same task with 0.5 offset (should not affect predictions)
-  data = task$data()
-  # see https://xgboost.readthedocs.io/en/stable/tutorials/intercept.html
-  set(data, j = "zeros", value = rep(0.5, nrow(data)))
-  task_offset = as_task_regr(data, target = "mpg")
+  # main task
+  task = as_task_regr(data, target = "y")
+
+  # task with a zero constant offset
+  task_offset = task$clone()
+  task_offset$cbind(data.frame(zeros = rep(0, n)))
   task_offset$set_col_roles(cols = "zeros", roles = "offset")
 
-  # same task but with a numeric column acting as offset
+  # task with a random offset per observation
   task_offset2 = task$clone()
-  task_offset2$set_col_roles(cols = "qsec", roles = "offset")
+  task_offset2$cbind(data.frame(random_offset = runif(n)))
+  task_offset2$set_col_roles(cols = "random_offset", roles = "offset")
 
   # add predefined internal validation task
   part = partition(task, c(0.6, 0.2)) # 60% train, 20% test, 20% validate
@@ -247,14 +256,76 @@ test_that("base_margin (offset)", {
   task_offset$internal_valid_task = part$validation
   task_offset2$internal_valid_task = part$validation
 
-  l = lrn("regr.xgboost", nrounds = 5)
-  l$validate = "predefined"
-  p1 = l$train(task, part$train)$predict(task, part$test) # no offset
-  p2 = l$train(task_offset, part$train)$predict(task_offset, part$test) # zero offset
-  expect_false("zeros" %in% l$model$feature_names) # offset column is not a feature
-  p3 = l$train(task_offset2, part$train)$predict(task_offset2, part$test) # non-zero offset
-  expect_false("qsec" %in% l$model$feature_names) # "qsec" column is not a feature
+  # use Poisson where base_score = 1 is equivalent to base_margin = 0:
+  # log(mean-scale) = base_margin + f(x)
+  # https://xgboost.readthedocs.io/en/stable/tutorials/intercept.html#offset
+  l0 = lrn("regr.xgboost", objective = "count:poisson",
+           nrounds = 3, base_score = 1)
+  l1 = lrn("regr.xgboost", objective = "count:poisson",
+           nrounds = 3) # estimated base_score (learned global intercept)
+  l2 = l1$clone(deep = TRUE) # train and test on task_offset
+  l3 = l1$clone(deep = TRUE) # train and test on task_offset2
+  l0$validate = "predefined"
+  l1$validate = "predefined"
+  l2$validate = "predefined"
+  l3$validate = "predefined"
 
-  #expect_equal(p1$response, p2$response) # zero offset => same predictions
-  expect_false(all(p1$response == p3$response)) # non-zero offset => different predictions
+  # train
+  l0$train(task, part$train) # uses base_score = 1
+  l1$train(task, part$train) # estimates base_score
+  l2$train(task_offset, part$train) # uses base_margin = 0
+  l3$train(task_offset2, part$train) # uses random base_margin
+  # if you peek inside the l1, l2 and l3 xgboost models, you will see a
+  # base_score value that has been estimated (i.e. to be used for prediction)
+  # https://github.com/dmlc/xgboost/issues/11872#issuecomment-3666848941
+
+  # different models: base_score = 1 vs estimated base_score
+  expect_false(length(xgboost::xgb.dump(l0$model)) == length(xgboost::xgb.dump(l1$model)))
+  # different models: estimated base_score vs base_margin = 0
+  expect_false(length(xgboost::xgb.dump(l1$model)) == length(xgboost::xgb.dump(l2$model)))
+  # different models: estimated base_score vs random base_margin
+  expect_false(length(xgboost::xgb.dump(l1$model)) == length(xgboost::xgb.dump(l3$model)))
+  # same models: base_score = 1 <=> base_margin = 0
+  expect_equal(xgboost::xgb.dump(l0$model), xgboost::xgb.dump(l2$model))
+
+  # l2 model: "zeros" is not a feature (it is the offset), "x" is a feature
+  expect_true("zeros" %nin% xgboost::getinfo(l2$model, "feature_name"))
+  expect_true("x" %in% xgboost::getinfo(l2$model, "feature_name"))
+  # similar for the l3 model
+  expect_true("random_offset" %nin% xgboost::getinfo(l3$model, "feature_name"))
+  expect_true("x" %in% xgboost::getinfo(l3$model, "feature_name"))
+
+  # predict (default: use_pred_offset = TRUE)
+  p0 = l0$predict(task, part$test) # task has no offset, base_score = 1 is used
+  p1 = l1$predict(task, part$test) # task has no offset, the estimated global intercept from train data is used
+  p2 = l2$predict(task_offset, part$test) # zero offset from test data applied
+  p3 = l3$predict(task_offset2, part$test) # random offset from test data is applied
+  # when different offsets are used, we expect different results
+  expect_false(all(p0$response == p1$response))
+  expect_false(all(p0$response == p3$response))
+  expect_false(all(p1$response == p2$response))
+  expect_false(all(p1$response == p3$response))
+  expect_false(all(p2$response == p3$response))
+  # predictions came from exactly the same models + the test offsets were exactly
+  # the same: base_score = 1 (p0) and base_margin = 0 (p2)
+  expect_equal(p0$response, p2$response)
+
+  # predict (use_pred_offset = FALSE)
+  l0$param_set$set_values(use_pred_offset = FALSE)
+  l1$param_set$set_values(use_pred_offset = FALSE)
+  l2$param_set$set_values(use_pred_offset = FALSE)
+  l3$param_set$set_values(use_pred_offset = FALSE)
+
+  # task has no offset => offset is taken from the train model as before
+  p00 = l0$predict(task, part$test)
+  p11 = l1$predict(task, part$test)
+  expect_equal(p0$response, p00$response) # base_score = 1 is used
+  expect_equal(p1$response, p11$response) # estimated offset from train data is used
+
+  # tasks have offsets now but they are disabled upon prediction!
+  p22 = l2$predict(task_offset, part$test)
+  p33 = l3$predict(task_offset2, part$test)
+  # these differ from previous predictions (where task offsets were applied)
+  expect_true(all(p2$response != p22$response))
+  expect_true(all(p3$response != p33$response))
 })